+++ src/OpenTelemetry.Api/Trace/ActivityExtensions.cs
@@ -15,9 +15,9 @@
 // </copyright>
 
 using System;
+using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
-using System.Linq.Expressions;
+using System.Reflection;
 using System.Runtime.CompilerServices;
 using OpenTelemetry.Internal;
 
@@ -28,6 +28,8 @@ namespace OpenTelemetry.Trace
     /// </summary>
     public static class ActivityExtensions
     {
+        private static readonly object EmptyActivityTagObjects = typeof(Activity).GetField("s_emptyTagObjects", BindingFlags.Static | BindingFlags.NonPublic).GetValue(null);
+
         /// <summary>
         /// Sets the status of activity execution.
         /// Activity class in .NET does not support 'Status'.
@@ -62,17 +64,51 @@ namespace OpenTelemetry.Trace
         {
             Debug.Assert(activity != null, "Activity should not be null");
 
-            var statusCanonicalCode = activity.Tags.FirstOrDefault(k => k.Key == SpanAttributeConstants.StatusCodeKey).Value;
-            var statusDescription = activity.Tags.FirstOrDefault(d => d.Key == SpanAttributeConstants.StatusDescriptionKey).Value;
+            ActivityStatusTagEnumerator state = default;
 
-            var status = SpanHelper.ResolveCanonicalCodeToStatus(statusCanonicalCode);
+            ActivityTagObjectsEnumeratorFactory<ActivityStatusTagEnumerator>.Enumerate(activity, ref state);
 
-            if (status.IsValid && !string.IsNullOrEmpty(statusDescription))
+            var status = SpanHelper.ResolveCanonicalCodeToStatus(state.StatusCode);
+
+            if (status.IsValid && !string.IsNullOrEmpty(state.StatusDescription))
             {
-                return status.WithDescription(statusDescription);
+                return status.WithDescription(state.StatusDescription);
             }
 
             return status;
+        }
+
+        /// <summary>
+        /// Gets the value of a specific tag on an <see cref="Activity"/>.
+        /// </summary>
+        /// <param name="activity">Activity instance.</param>
+        /// <param name="tagName">Case-sensitive tag name to retrieve.</param>
+        /// <returns>Tag value or null if a match was not found.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static object GetTagValue(this Activity activity, string tagName)
+        {
+            Debug.Assert(activity != null, "Activity should not be null");
+
+            ActivitySingleTagEnumerator state = new ActivitySingleTagEnumerator(tagName);
+
+            ActivityTagObjectsEnumeratorFactory<ActivitySingleTagEnumerator>.Enumerate(activity, ref state);
+
+            return state.Value;
+        }
+
+        /// <summary>
+        /// Enumerates all the key/value pairs on an <see cref="Activity"/> without performing an allocation.
+        /// </summary>
+        /// <typeparam name="T">The struct <see cref="IActivityTagEnumerator"/> implementation to use for the enumeration.</typeparam>
+        /// <param name="activity">Activity instance.</param>
+        /// <param name="tagEnumerator">Tag enumerator.</param>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static void EnumerateTagValues<T>(this Activity activity, ref T tagEnumerator)
+            where T : struct, IActivityTagEnumerator
+        {
+            Debug.Assert(activity != null, "Activity should not be null");
+
+            ActivityTagObjectsEnumeratorFactory<T>.Enumerate(activity, ref tagEnumerator);
         }
 
         /// <summary>
@@ -100,6 +136,81 @@ namespace OpenTelemetry.Trace
             }
 
             activity?.AddEvent(new ActivityEvent(SemanticConventions.AttributeExceptionEventName, default, tagsCollection));
+        }
+
+        private struct ActivitySingleTagEnumerator : IActivityTagEnumerator
+        {
+            private readonly string tagName;
+
+            public ActivitySingleTagEnumerator(string tagName)
+            {
+                this.tagName = tagName;
+                this.Value = null;
+            }
+
+            public object Value { get; private set; }
+
+            public bool ForEach(KeyValuePair<string, object> item)
+            {
+                if (item.Key == this.tagName)
+                {
+                    this.Value = item.Value;
+                    return false;
+                }
+
+                return true;
+            }
+        }
+
+        private struct ActivityStatusTagEnumerator : IActivityTagEnumerator
+        {
+            public string StatusCode { get; private set; }
+
+            public string StatusDescription { get; private set; }
+
+            public bool ForEach(KeyValuePair<string, object> item)
+            {
+                switch (item.Key)
+                {
+                    case SpanAttributeConstants.StatusCodeKey:
+                        this.StatusCode = item.Value as string;
+                        break;
+                    case SpanAttributeConstants.StatusDescriptionKey:
+                        this.StatusDescription = item.Value as string;
+                        break;
+                }
+
+                return this.StatusCode == null || this.StatusDescription == null;
+            }
+        }
+
+        private static class ActivityTagObjectsEnumeratorFactory<TState>
+            where TState : struct, IActivityTagEnumerator
+        {
+            private static readonly DictionaryEnumerator<string, object, TState>.AllocationFreeForEachDelegate
+                ActivityTagObjectsEnumerator = DictionaryEnumerator<string, object, TState>.BuildAllocationFreeForEachDelegate(
+                    typeof(Activity).GetField("_tags", BindingFlags.Instance | BindingFlags.NonPublic).FieldType);
+
+            private static readonly DictionaryEnumerator<string, object, TState>.ForEachDelegate ForEachTagValueCallbackRef = ForEachTagValueCallback;
+
+            [MethodImpl(MethodImplOptions.AggressiveInlining)]
+            public static void Enumerate(Activity activity, ref TState state)
+            {
+                var tagObjects = activity.TagObjects;
+
+                if (ReferenceEquals(tagObjects, EmptyActivityTagObjects))
+                {
+                    return;
+                }
+
+                ActivityTagObjectsEnumerator(
+                    tagObjects,
+                    ref state,
+                    ForEachTagValueCallbackRef);
+            }
+
+            private static bool ForEachTagValueCallback(ref TState state, KeyValuePair<string, object> item)
+                => state.ForEach(item);
         }
     }
 }