+++ src/OpenTelemetry.Instrumentation.GrpcNetClient/Implementation/GrpcClientDiagnosticListener.cs
@@ -15,7 +15,6 @@
 // </copyright>
 using System;
 using System.Diagnostics;
-using System.Globalization;
 using System.Net.Http;
 using OpenTelemetry.Trace;
 
@@ -24,11 +23,12 @@ namespace OpenTelemetry.Instrumentation.GrpcNetClient.Implementation
     internal class GrpcClientDiagnosticListener : ListenerHandler
     {
         public const string RequestCustomPropertyName = "OTel.GrpcHandler.Request";
+        private readonly GrpcClientInstrumentationOptions options;
 
         private readonly ActivitySourceAdapter activitySource;
-        private readonly PropertyFetcher startRequestFetcher = new PropertyFetcher("Request");
+        private readonly PropertyFetcher<HttpRequestMessage> startRequestFetcher = new PropertyFetcher<HttpRequestMessage>("Request");
 
-        public GrpcClientDiagnosticListener(ActivitySourceAdapter activitySource)
+        public GrpcClientDiagnosticListener(ActivitySourceAdapter activitySource, GrpcClientInstrumentationOptions options)
             : base("Grpc.Net.Client")
         {
             if (activitySource == null)
@@ -36,6 +36,7 @@ namespace OpenTelemetry.Instrumentation.GrpcNetClient.Implementation
                 throw new ArgumentNullException(nameof(activitySource));
             }
 
+            this.options = options;
             this.activitySource = activitySource;
         }
 
@@ -45,6 +46,11 @@ namespace OpenTelemetry.Instrumentation.GrpcNetClient.Implementation
             {
                 GrpcInstrumentationEventSource.Log.NullPayload(nameof(GrpcClientDiagnosticListener), nameof(this.OnStartActivity));
                 return;
+            }
+
+            if (this.options.SuppressDownstreamInstrumentation)
+            {
+                SuppressInstrumentationScope.Enter();
             }
 
             var grpcMethod = GrpcTagHelper.GetGrpcMethodFromActivity(activity);
@@ -62,6 +68,9 @@ namespace OpenTelemetry.Instrumentation.GrpcNetClient.Implementation
                 {
                     activity.SetTag(SemanticConventions.AttributeRpcService, rpcService);
                     activity.SetTag(SemanticConventions.AttributeRpcMethod, rpcMethod);
+
+                    // Remove the grpc.method tag added by the gRPC .NET library
+                    activity.SetTag(GrpcTagHelper.GrpcMethodTagName, null);
                 }
 
                 var uriHostNameType = Uri.CheckHostName(request.RequestUri.Host);
@@ -74,7 +83,7 @@ namespace OpenTelemetry.Instrumentation.GrpcNetClient.Implementation
                     activity.SetTag(SemanticConventions.AttributeNetPeerName, request.RequestUri.Host);
                 }
 
-                activity.SetTag(SemanticConventions.AttributeNetPeerPort, request.RequestUri.Port.ToString(CultureInfo.InvariantCulture));
+                activity.SetTag(SemanticConventions.AttributeNetPeerPort, request.RequestUri.Port);
             }
         }
 
@@ -83,6 +92,9 @@ namespace OpenTelemetry.Instrumentation.GrpcNetClient.Implementation
             if (activity.IsAllDataRequested)
             {
                 activity.SetStatus(GrpcTagHelper.GetGrpcStatusCodeFromActivity(activity));
+
+                // Remove the grpc.status_code tag added by the gRPC .NET library
+                activity.SetTag(GrpcTagHelper.GrpcStatusCodeTagName, null);
             }
 
             this.activitySource.Stop(activity);