+++ src/OpenTelemetry/CHANGELOG.md
@@ -2,7 +2,15 @@
 
 ## Unreleased
 
-## 0.5.0-beta.1
+* Fixes [953](https://github.com/open-telemetry/opentelemetry-dotnet/issues/953)
+* Changes arising from `DiagnosticSource` changes
+  ([#1203](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1203))
+* `PropertyFetcher` is now public
+  ([#1232](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1232))
+* `PropertyFetcher` changed to `PropertyFetcher<T>`
+  ([#1238](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1238))
+
+## 0.5.0-beta.2
 
 Released 2020-08-28